@@ -6,15 +6,24 @@
   "description": "Let's try write js-Game",
   "main": "src/server/index.js",
   "dependencies": {
+    "@types/react": "^16.4.7",
+    "@types/react-dom": "^16.0.6",
+    "@types/ws": "^5.1.2",
     "body-parser": "latest",
     "express": "^4.16.3",
+    "inversify": "^4.13.0",
     "mongodb": "^3.1.1",
-    "react": "^16.3.1",
-    "react-dom": "^16.3.1",
-    "uuid": "^3.3.2",
+    "mongoose": "^5.2.6",
+    "react": "^16.4.1",
+    "react-dom": "^16.4.1",
+    "reflect-metadata": "^0.1.12",
+    "ts-loader": "^4.4.2",
     "ws": "^5.2.2"
   },
   "devDependencies": {
+    "@types/express": "^4.16.0",
+    "@types/mongodb": "^3.1.2",
+    "@types/mongoose": "^5.2.3",
     "babel-core": "^6.26.0",
     "babel-loader": "^7.1.4",
     "babel-plugin-transform-class-properties": "^6.24.1",
@@ -32,6 +41,7 @@
     "html-webpack-plugin": "^3.2.0",
     "nodemon": "^1.18.3",
     "style-loader": "^0.21.0",
+    "typescript": "^3.0.1",
     "webpack": "^4.16.3",
     "webpack-cli": "^3.1.0",
     "webpack-dev-server": "^3.1.5"
@@ -39,8 +49,8 @@
   "scripts": {
     "dev": "concurrently \"npm run dev:server\" \"npm run dev:client\"",
     "dev:client": "webpack-dev-server --mode development --devtool inline-source-map --hot",
-    "dev:server": "nodemon src/server/index.js",
+    "dev:server": "nodemon src/server/index.ts",
     "prod:build": "webpack --mode production",
-    "prod:start": "node src/server/index.js"
+    "prod:start": "node src/server/index.ts"
   }
 }
