@@ -2,4 +2,4 @@ require('babel-register')({
     presets: ['env']
 });
 
-module.exports = require('./server.js');
+module.exports = require('./server.ts');
